refactor(client): rename the utilities directory
