+++ packages/client/src/apollo.ts
@@ -7,7 +7,7 @@ import {
   createErrorLink,
   createFragmentMatcher,
   createInMemoryCache,
-} from './utilities/apollo'
+} from './utils/apollo'
 
 interface Options {
   defaults: {}
