+++ packages/client/src/components/phone-field/input.tsx
@@ -2,7 +2,7 @@ import { FieldValidator, useField, useFormikContext } from 'formik'
 import React from 'react'
 
 import BaseInput, { Props as InputProps } from 'components/input'
-import { validateInput } from 'utilities/validation'
+import { validateInput } from 'utils/validation'
 import { usePhoneFieldContext } from '.'
 
 interface Props
