+++ packages/client/src/routes/home/stories.tsx
@@ -2,7 +2,7 @@ import { MockedProvider } from '@apollo/react-testing'
 import { storiesOf } from '@storybook/react'
 import React from 'react'
 
-import { createFragmentMatcher, createInMemoryCache } from 'utilities/apollo'
+import { createFragmentMatcher, createInMemoryCache } from 'utils/apollo'
 import Home from '.'
 
 const fragmentMatcher = createFragmentMatcher()
