+++ packages/client/src/routes/recover/stories.tsx
@@ -2,13 +2,13 @@ import { MockedProvider } from '@apollo/react-testing'
 import { storiesOf } from '@storybook/react'
 import React from 'react'
 
-import { createFragmentMatcher, createInMemoryCache } from 'utilities/apollo'
+import { createFragmentMatcher, createInMemoryCache } from 'utils/apollo'
 import {
   emailBuilder,
   profileBuilder,
   responseBuilder,
   userBuilder,
-} from 'utilities/specification'
+} from 'utils/specification'
 import { currentUserQuery, recoverMutation } from './graphql'
 import Recover from '.'
