+++ packages/client/src/utils/validation.ts
